Correcting byte issue.

+++ DHLWebAPI/Models/DTOs/TblUsersDTO.cs
@@ -9,7 +9,7 @@ namespace DHLWebAPI.Models.DTOs
     {
         public int IdUser { get; set; }
         public string ContactName { get; set; }
-        public byte UserType { get; set; }
+        public int UserType { get; set; }
         public string TitleOfCourtesy { get; set; }
         public DateTime Birthdate { get; set; }
         public DateTime Hiredate { get; set; }
